# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "resturan",
        "title": "Gin Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v2/foodgroup": {
            "get": {
                "description": "Get all food group",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "food group"
                ],
                "summary": "Get all food group",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pagecount",
                        "name": "pagecount",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "description": "Update a food group",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "food group"
                ],
                "summary": "Update a food group",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "food group id",
                        "name": "Id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "food group name",
                        "name": "Name",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "food group image",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "create a food group",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "food group"
                ],
                "summary": "create a food group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "food group name",
                        "name": "Name",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "food group image",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}